extract utils method
