+++ packages/code-du-travail-modeles/src/modeles/common/utils/date.ts
@@ -1,6 +1,6 @@
 import { parse } from "date-fns";
 
-import type { SalaryPeriods } from "..";
+import type { Absence, SalaryPeriods } from "..";
 
 /**
  *
@@ -126,4 +126,50 @@ export const monthToNumber = (month: string): number => {
     default:
       return 0;
   }
+};
+
+export const calculateDurationByYear = (
+  absencePeriods: Absence[]
+): number[] => {
+  const absencePeriodsWithStartDate: {
+    startedAtDate: Date;
+    durationInMonth: number;
+  }[] = absencePeriods
+    .filter((absencePeriod) => !!absencePeriod.startedAt)
+    .filter((absencePeriod) => !!absencePeriod.durationInMonth)
+    .map((absencePeriod) => {
+      return {
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        durationInMonth: absencePeriod.durationInMonth!,
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        startedAtDate: parseDate(absencePeriod.startedAt!),
+      };
+    });
+
+  const durationByYear = absencePeriodsWithStartDate.reduce(
+    (total: Record<number, number>, abs) => {
+      const startYear = abs.startedAtDate.getFullYear();
+      let startMonth = abs.startedAtDate.getMonth();
+      let remainingDuration = abs.durationInMonth;
+
+      for (let year = startYear; remainingDuration > 0; year++) {
+        const remainingMonthsInYear = 12 - startMonth;
+        const durationInCurrentYear = Math.min(
+          remainingDuration,
+          remainingMonthsInYear
+        );
+
+        // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+        total[year] = (total[year] ?? 0) + durationInCurrentYear;
+
+        remainingDuration -= durationInCurrentYear;
+        startMonth = 0;
+      }
+
+      return total;
+    },
+    {}
+  );
+
+  return Object.values(durationByYear);
 };
